add font-family to generic input styles (#3040)
